requirements.txt
requests~=2.31.0
web3
python-telegram-bot==13.15
python-dotenv~=1.0.1